refactor interpolators
